fix: use correct unsigned integer types for MySQL columns
sqlx 0.8.x enforces strict type matching between Rust and MySQL types.
Changed i32 to u32 for columns defined as INT UNSIGNED in the database:
- archive.rs: year, month (from YEAR/MONTH functions), count to i64
- canteen_pivot_menus_data.rs: data_id, menu_id
- posts_pivot_labels_data.rs: labels_id, posts_id
- menu_items.rs: lft, rgt, depth

+++ src/graphql/resolvers/archive.rs
@@ -17,9 +17,9 @@ use std::{ops::Deref, sync::Arc};
 
 #[derive(SimpleObject, Debug, FromQueryResult)]
 pub struct Info {
-    count: i32,
-    year: i32,
-    month: i32,
+    count: i64,
+    year: u32,
+    month: u32,
 }
 
 #[derive(Debug)]
